ViesteToken
welcome to the pre-sale of ViesteToken-crypto.com, the only token to be created for vieste. ViesteToken is the official token of the city of Vieste, it was created to make our city known to the crypto world, furthermore Vieste is a very important tourist destination in the Apulian region in Italy. ViesteToken-crypto.com represents our idea of ​​being the first village in the world to have a personal cryptocurrency. this is the cryptocurrency of vieste, visit our website Viestetoken-crypto.com support the project.